requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
cryptography==41.0.3
python-dotenv==1.0.1
PyJWT==2.8.0
numpy==1.26.2
reportlab==4.0.7